Move stylesheets related code to DocumentOrShadowRoot
